refactor: Phase 3 slice 3d — shikimori router + ShikimoriSyncService (#84)
Fourth slice (3d) of Phase 3 — migrates the 16 Shikimori IPC handlers
behind the router pattern and extracts a ShikimoriSyncService that owns
the sync timer, offline-queue drain worker, anime-details prefetcher,
and calendar cache. Behavior-identical, no IPC surface change.

* New src/main/services/shikimori-sync/index.ts —
  createShikimoriSyncService. Encapsulates all sync/prefetch/calendar
  module-level state that previously lived as free vars in index.ts
  (syncInProgress, syncTimer, lastSyncAt/Error, prefetchInProgress/
  Abort, calendarCache). Exposes syncShikimoriQueue (consolidate →
  drift-resolve → reconcile), prefetchShikimoriDetails,
  refreshShikimoriDetailsForMalId, the sync-timer trio
  (start/stop/adjust), getSyncStatus/getQueueLength,
  broadcastSyncStatus, abortPrefetch, and the calendar-cache
  accessors. Deps-injected (store, broadcast + 4 channel names).

* New src/main/ipc/shikimori.ipc.ts — all 16 CHANNELS.SHIKIMORI_*
  handlers: OAuth (GET_AUTH_URL/EXCHANGE_CODE/LOGOUT/GET_USER), rate
  read/write (GET_RATE, UPDATE_RATE with the offline-queue fallback
  on network error), GET_FRIENDS_RATES, GET_ANIME_RATES (cache-first
  + background refresh), GET_ANIME_DETAILS, TRIGGER_DETAIL_PREFETCH,
  GET_OFFLINE_QUEUE_LENGTH, GET_SYNC_STATUS, TRIGGER_SYNC,
  GET_FRIENDS_ACTIVITY, GET_CALENDAR, GET_RELATED. The
  fetchAndCacheShikimoriRates / refreshShikimoriRatesInBackground
  orchestration moves into the router as local functions.

* AppDeps grows: shikimoriSyncService, lookupByMalIds,
  maybeBroadcastCleanupPrompt, runAutoDownloadTick, and a generic
  broadcast fan-out (the index.ts broadcastToAll).

* src/main/index.ts removes ~350 LOC of sync helpers + ~465 LOC of
  handlers; builds the service after coldStorageService; rewires the
  two external call sites (initAutoDownloader.refreshShikimoriDetails
  and the boot-time queue-drain check) to the service; drops the
  now-unused isNetworkError / consolidateQueue / QueuedShikimoriUpdate
  imports.

* registerIpcHandlers() down to ~55 handlers (settings,
  watch-progress, skip-detector, storage, file/shell, player,
  syncplay, auto-download, debug) — those land in 3e–3g.

* DESIGN.md adds rows for the new service + router.

Stacked on phase3-pr3-downloads-cleanup; refs #102, #84.
