+++ DESIGN.md
@@ -40,6 +40,8 @@ Renderer (Vue)  --ipcRenderer.invoke-->  Preload (bridge)  --ipcMain.handle-->  
 | `src/main/ipc/ffmpeg.ipc.ts` | Ffmpeg router: `CHANNELS.{FFMPEG_CHECK, FFMPEG_DELETE}` — checks via the injected `checkFfmpeg`, deletes binaries + `ffbinaries-cache` zip cache, and clears the mutable `ffmpegPath`/`ffprobePath` module variables via `clearFfmpegPaths()` |
 | `src/main/ipc/cleanup.ipc.ts` | Cleanup router: `CHANNELS.{CLEANUP_GET_SIZE, CLEANUP_GET_ACTIVE_DOWNLOADS, CLEANUP_EXECUTE, CLEANUP_GET_SNOOZED, CLEANUP_SET_SNOOZED}` — uses `coldStorageService.dirsForScan()` to delete the show across hot+cold dirs, prunes `downloadedEpisodes` keys, drops anime/skip-detection cache entries, snoozes per anime id |
 | `src/main/ipc/downloaded-anime.ipc.ts` | Downloaded-anime router: `CHANNELS.{DOWNLOADED_ANIME_ADD, DOWNLOADED_ANIME_DELETE}` — delete cascades to all storage dirs, anime-cache entry, skip-detection cache + detections |
+| `src/main/ipc/shikimori.ipc.ts` | Shikimori router (Phase 3 slice 3d): all 16 `CHANNELS.SHIKIMORI_*` handlers — OAuth (`GET_AUTH_URL`/`EXCHANGE_CODE`/`LOGOUT`/`GET_USER`), rate read/write (`GET_RATE`/`UPDATE_RATE` with offline-queue fallback on network error), `GET_FRIENDS_RATES`/`GET_ANIME_RATES` (cache-first + background refresh)/`GET_ANIME_DETAILS`/`TRIGGER_DETAIL_PREFETCH`, sync surface (`GET_OFFLINE_QUEUE_LENGTH`/`GET_SYNC_STATUS`/`TRIGGER_SYNC`), `GET_FRIENDS_ACTIVITY`/`GET_CALENDAR`/`GET_RELATED`. Holds the `fetchAndCacheShikimoriRates`/`refreshShikimoriRatesInBackground` orchestration locally |
+| `src/main/services/shikimori-sync/index.ts` | `createShikimoriSyncService` — owns the Shikimori sync timer + offline-queue drain worker (`syncShikimoriQueue`, consolidate → drift-resolve → reconcile), the anime-details prefetcher (`prefetchShikimoriDetails`/`refreshShikimoriDetailsForMalId`), and the 5-minute calendar cache. All sync/prefetch/calendar module state (formerly free vars in `index.ts`) now lives in the service closure. Deps-injected (`store`, `broadcast` + 4 channel names) |
 | `src/main/smotret-api.ts` | Smotret-Anime API client: search, anime/episode details, embed, subtitles, token validation |
 | `src/main/download-manager.ts` | Download queue, concurrent downloads, progress, ffmpeg merge, scan-merge |
 | `src/main/shikimori.ts` | Shikimori API client: OAuth, token refresh, user rates, anime list |
