+++ src/main/index.ts
@@ -26,8 +26,6 @@ import { pathToFileURL } from 'url'
 import { Readable } from 'stream'
 import { randomUUID } from 'crypto'
 import { SmotretApi } from './smotret-api'
-import { isNetworkError } from './lib/errors'
-import { consolidateQueue, type QueuedShikimoriUpdate } from './lib/shikimori-queue'
 import { parseEpisodeFromFilename } from './lib/filename'
 import { avcCodecString, hevcCodecString, aacCodecString } from './streaming/codec-strings'
 import type { AnimeSearchResult, AnimeDetail, Translation } from './smotret-api'
@@ -61,6 +59,7 @@ import type { Mp4StreamingStats } from './mp4-faststart'
 import { createAnimeCacheService, type AnimeCacheEntry } from './services/anime-cache'
 import { createSkipAnalysisService } from './services/skip-analysis'
 import { createColdStorageService } from './services/cold-storage'
+import { createShikimoriSyncService } from './services/shikimori-sync'
 import { createStreamingService, type MseSession, type MseOpenResult } from './streaming'
 import { App } from './app/core'
 import { registerIpcRouters } from './ipc'
@@ -243,358 +242,6 @@ async function recordMp4FaststartCheck(
   })
   mp4StatsWriteChain = next.catch(() => undefined)
   await next
-}
-
-type ShikiSyncState = 'idle' | 'syncing'
-interface ShikiSyncStatus {
-  state: ShikiSyncState
-  queueLength: number
-  lastSyncAt: number
-  lastSyncError: string | null
-}
-
-const SYNC_TIMER_MS = 60_000
-const SYNC_ITEM_DELAY_MS = 250
-
-const STATUS_ORDER: Record<shikimori.ShikiUserRateStatus, number> = {
-  planned: 0,
-  watching: 1,
-  rewatching: 1,
-  on_hold: 2,
-  dropped: 2,
-  completed: 3
-}
-
-let syncInProgress = false
-let syncTimer: NodeJS.Timeout | null = null
-let lastSyncAt = 0
-let lastSyncError: string | null = null
-
-const PREFETCH_INTER_REQUEST_MS = 2000
-const PREFETCH_STALENESS_MS = 30 * 24 * 60 * 60 * 1000
-const PREFETCH_STATUSES = new Set<shikimori.ShikiUserRateStatus>(['watching', 'planned'])
-
-let prefetchInProgress = false
-let prefetchAbort = false
-
-const CALENDAR_CACHE_TTL_MS = 5 * 60 * 1000
-let calendarCache: { data: CalendarEntry[]; fetchedAt: number } | null = null
-function invalidateCalendarCache(): void {
-  calendarCache = null
-}
-
-function getQueueLength(): number {
-  return (store.get('shikimoriUpdateQueue') as QueuedShikimoriUpdate[]).length
-}
-
-function getSyncStatus(): ShikiSyncStatus {
-  return {
-    state: syncInProgress ? 'syncing' : 'idle',
-    queueLength: getQueueLength(),
-    lastSyncAt,
-    lastSyncError
-  }
-}
-
-function broadcastSyncStatus(): void {
-  broadcastToAll(EVENT_CHANNELS.SHIKIMORI_SYNC_STATUS, getSyncStatus())
-}
-
-function startSyncTimer(): void {
-  if (syncTimer) return
-  syncTimer = setInterval(() => {
-    void syncShikimoriQueue()
-  }, SYNC_TIMER_MS)
-}
-
-function stopSyncTimer(): void {
-  if (!syncTimer) return
-  clearInterval(syncTimer)
-  syncTimer = null
-}
-
-function adjustSyncTimer(): void {
-  if (getQueueLength() > 0) startSyncTimer()
-  else stopSyncTimer()
-}
-
-function dropConsumedEntries(malId: number, consumedQueuedAts: number[]): number {
-  const consumed = new Set(consumedQueuedAts)
-  const queue = (store.get('shikimoriUpdateQueue') as QueuedShikimoriUpdate[]).filter(
-    (q) => !(q.malId === malId && consumed.has(q.queuedAt))
-  )
-  store.set('shikimoriUpdateQueue', queue)
-  broadcastToAll(EVENT_CHANNELS.SHIKIMORI_OFFLINE_QUEUE_CHANGED, { length: queue.length })
-  return queue.length
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms))
-}
-
-function reconcileCacheFromRate(malId: number, rate: shikimori.ShikiUserRate): void {
-  const cached = store.get('shikimoriUserRates') as {
-    rate: Record<string, unknown> & {
-      id?: number
-      target_id: number
-      episodes: number
-      status: shikimori.ShikiUserRateStatus
-      score: number
-      rewatches?: number
-    }
-    shikiAnime: unknown
-    smotretAnime: unknown
-  }[]
-  const idx = cached.findIndex((e) => e.rate.target_id === malId)
-  if (idx === -1) return
-  cached[idx] = {
-    ...cached[idx],
-    rate: {
-      ...cached[idx].rate,
-      id: rate.id,
-      episodes: rate.episodes,
-      status: rate.status,
-      score: rate.score,
-      rewatches: rate.rewatches ?? 0,
-      updated_at: new Date().toISOString()
-    }
-  }
-  store.set('shikimoriUserRates', cached)
-  invalidateCalendarCache()
-  broadcastToAll(EVENT_CHANNELS.SHIKIMORI_RATE_UPDATED, cached[idx])
-}
-
-function shouldApplyOnDrift(
-  current: shikimori.ShikiUserRate,
-  after: { episodes: number; status: shikimori.ShikiUserRateStatus; score: number }
-): boolean {
-  if (after.episodes <= current.episodes) return false
-  const currentRank = STATUS_ORDER[current.status]
-  const afterRank = STATUS_ORDER[after.status]
-  if (currentRank > afterRank) return false
-  // Same rank but different value (watching ↔ rewatching) is a user-meaningful side-grade,
-  // not pure episode progress — treat as drift and let the server value win
-  if (currentRank === afterRank && current.status !== after.status) return false
-  return true
-}
-
-async function syncShikimoriQueue(): Promise<void> {
-  if (syncInProgress) return
-  const queue = store.get('shikimoriUpdateQueue') as QueuedShikimoriUpdate[]
-  if (queue.length === 0) {
-    stopSyncTimer()
-    return
-  }
-  const creds = store.get('shikimoriCredentials') as shikimori.ShikiCredentials | null
-  const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-  if (!creds || !user) return
-
-  syncInProgress = true
-  broadcastSyncStatus()
-
-  let accessToken: string
-  try {
-    accessToken = await shikimori.ensureFreshToken(store)
-  } catch (err) {
-    syncInProgress = false
-    if (!isNetworkError(err)) {
-      lastSyncError = err instanceof Error ? err.message : String(err)
-      console.warn('[shikimori sync] auth refresh failed:', err)
-    }
-    broadcastSyncStatus()
-    return
-  }
-
-  const work = consolidateQueue(queue)
-  let aborted = false
-
-  for (const item of work) {
-    try {
-      const current = await shikimori.getUserRate(accessToken, user.id, item.malId)
-
-      if (!current) {
-        const created = await shikimori.createUserRate(
-          accessToken,
-          user.id,
-          item.malId,
-          item.after.episodes,
-          item.after.status,
-          item.after.score,
-          item.after.rewatches
-        )
-        reconcileCacheFromRate(item.malId, created)
-      } else {
-        const driftMatches =
-          current.episodes === item.before.episodes &&
-          current.status === item.before.status &&
-          current.score === item.before.score
-
-        if (driftMatches) {
-          const updated = await shikimori.updateUserRate(
-            accessToken,
-            current.id,
-            item.after.episodes,
-            item.after.status,
-            item.after.score,
-            item.after.rewatches
-          )
-          reconcileCacheFromRate(item.malId, updated)
-        } else if (shouldApplyOnDrift(current, item.after)) {
-          const updated = await shikimori.updateUserRate(
-            accessToken,
-            current.id,
-            item.after.episodes,
-            item.after.status,
-            item.after.score,
-            item.after.rewatches
-          )
-          reconcileCacheFromRate(item.malId, updated)
-        } else {
-          reconcileCacheFromRate(item.malId, current)
-        }
-      }
-
-      dropConsumedEntries(item.malId, item.consumedQueuedAts)
-    } catch (err) {
-      if (isNetworkError(err)) {
-        aborted = true
-        break
-      }
-      if (err instanceof shikimori.ShikiApiError && (err.status === 401 || err.status === 403)) {
-        aborted = true
-        lastSyncError = err.message
-        console.warn('[shikimori sync] auth error, stopping drain:', err)
-        break
-      }
-      console.warn('[shikimori sync] dropping item', item.malId, err)
-      dropConsumedEntries(item.malId, item.consumedQueuedAts)
-    }
-
-    if (work.indexOf(item) < work.length - 1) await sleep(SYNC_ITEM_DELAY_MS)
-  }
-
-  syncInProgress = false
-  if (!aborted) {
-    lastSyncAt = Date.now()
-    lastSyncError = null
-  }
-  adjustSyncTimer()
-  broadcastSyncStatus()
-}
-
-function getStaleOrMissingMalIds(): number[] {
-  const rates = store.get('shikimoriUserRates') as {
-    rate: { target_id?: number; status: shikimori.ShikiUserRateStatus }
-    shikiAnime?: { id?: number }
-  }[]
-  const cache = store.get('shikimoriAnimeDetails') as Record<
-    string,
-    { details: shikimori.ShikiAnimeDetails; fetchedAt: number }
-  >
-  const now = Date.now()
-  const result: number[] = []
-  for (const entry of rates) {
-    if (!PREFETCH_STATUSES.has(entry.rate.status)) continue
-    const malId = entry.rate.target_id ?? entry.shikiAnime?.id
-    if (typeof malId !== 'number' || !Number.isFinite(malId) || malId <= 0) continue
-    const cached = cache[String(malId)]
-    if (!cached || now - cached.fetchedAt > PREFETCH_STALENESS_MS) {
-      result.push(malId)
-    }
-  }
-  return result
-}
-
-async function refreshShikimoriDetailsForMalId(
-  malId: number
-): Promise<shikimori.ShikiAnimeDetails | null> {
-  const creds = store.get('shikimoriCredentials') as shikimori.ShikiCredentials | null
-  const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-  if (!creds || !user) return null
-  let accessToken: string
-  try {
-    accessToken = await shikimori.ensureFreshToken(store)
-  } catch (err) {
-    if (!isNetworkError(err)) {
-      console.warn('[shikimori] single-show refresh auth failed:', err)
-    }
-    return null
-  }
-  try {
-    const details = await shikimori.getAnimeDetails(accessToken, malId)
-    const cache = store.get('shikimoriAnimeDetails') as Record<
-      string,
-      { details: shikimori.ShikiAnimeDetails; fetchedAt: number }
-    >
-    cache[String(malId)] = { details, fetchedAt: Date.now() }
-    store.set('shikimoriAnimeDetails', cache)
-    broadcastToAll(EVENT_CHANNELS.SHIKIMORI_ANIME_DETAILS_UPDATED, { malId, details })
-    return details
-  } catch (err) {
-    if (!isNetworkError(err)) {
-      console.warn('[shikimori] single-show refresh failed for', malId, err)
-    }
-    return null
-  }
-}
-
-async function prefetchShikimoriDetails(): Promise<void> {
-  if (prefetchInProgress) return
-  const creds = store.get('shikimoriCredentials') as shikimori.ShikiCredentials | null
-  const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-  if (!creds || !user) return
-
-  const work = getStaleOrMissingMalIds()
-  if (work.length === 0) return
-
-  prefetchInProgress = true
-  prefetchAbort = false
-
-  try {
-    let accessToken: string
-    try {
-      accessToken = await shikimori.ensureFreshToken(store)
-    } catch (err) {
-      if (!isNetworkError(err)) {
-        console.warn('[shikimori prefetch] auth refresh failed:', err)
-      }
-      return
-    }
-
-    for (let i = 0; i < work.length; i++) {
-      if (prefetchAbort) break
-      const malId = work[i]
-      try {
-        const details = await shikimori.getAnimeDetails(accessToken, malId)
-        const cache = store.get('shikimoriAnimeDetails') as Record<
-          string,
-          { details: shikimori.ShikiAnimeDetails; fetchedAt: number }
-        >
-        cache[String(malId)] = { details, fetchedAt: Date.now() }
-        store.set('shikimoriAnimeDetails', cache)
-        broadcastToAll(EVENT_CHANNELS.SHIKIMORI_ANIME_DETAILS_UPDATED, { malId, details })
-      } catch (err) {
-        if (isNetworkError(err)) {
-          console.warn('[shikimori prefetch] network error, aborting loop:', err)
-          break
-        }
-        if (err instanceof shikimori.ShikiApiError && (err.status === 401 || err.status === 403)) {
-          console.warn('[shikimori prefetch] auth error, aborting loop:', err)
-          break
-        }
-        if (err instanceof shikimori.ShikiApiError && err.status === 404) {
-          console.warn('[shikimori prefetch] anime', malId, 'not found, skipping')
-        } else {
-          console.warn('[shikimori prefetch] error fetching', malId, err)
-        }
-      }
-      if (i < work.length - 1) {
-        await sleep(PREFETCH_INTER_REQUEST_MS)
-      }
-    }
-  } finally {
-    prefetchInProgress = false
-  }
 }
 
 // --- Anime cache helpers (for offline support + fast-loading detail view) ---
@@ -892,6 +539,15 @@ const coldStorageService = createColdStorageService({
   cleanupPendingChannel: EVENT_CHANNELS.STORAGE_CLEANUP_PENDING,
   cleanupFinishedChannel: EVENT_CHANNELS.STORAGE_CLEANUP_FINISHED,
   fileEpisodesChangedChannel: EVENT_CHANNELS.FILE_EPISODES_CHANGED
+})
+
+const shikimoriSyncService = createShikimoriSyncService({
+  store,
+  broadcast: broadcastToAll,
+  syncStatusChannel: EVENT_CHANNELS.SHIKIMORI_SYNC_STATUS,
+  offlineQueueChangedChannel: EVENT_CHANNELS.SHIKIMORI_OFFLINE_QUEUE_CHANGED,
+  rateUpdatedChannel: EVENT_CHANNELS.SHIKIMORI_RATE_UPDATED,
+  animeDetailsUpdatedChannel: EVENT_CHANNELS.SHIKIMORI_ANIME_DETAILS_UPDATED
 })
 let downloadManager: DownloadManager
 let ffmpegPath = ''
@@ -1338,7 +994,12 @@ function registerIpcHandlers(): void {
     getFfmpegDir,
     sumShowFiles,
     checkFfmpeg,
-    getDisplayName
+    getDisplayName,
+    shikimoriSyncService,
+    lookupByMalIds,
+    maybeBroadcastCleanupPrompt,
+    runAutoDownloadTick,
+    broadcast: broadcastToAll
   })
 
   const qualityMismatches = new Map<
@@ -1776,472 +1437,6 @@ function registerIpcHandlers(): void {
 
   ipcMain.handle(CHANNELS.STORAGE_RUN_CLEANUP, async (_event, opts?: { force?: boolean }) => {
     return coldStorageService.runWatchedCleanup(!!opts?.force)
-  })
-
-  // Shikimori integration
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_AUTH_URL, () => {
-    return shikimori.getAuthUrl()
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_EXCHANGE_CODE, async (_event, code: string) => {
-    const creds = await shikimori.exchangeCode(code)
-    store.set('shikimoriCredentials', creds)
-    const user = await shikimori.getUser(creds.access_token)
-    store.set('shikimoriUser', user)
-    return user
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_LOGOUT, () => {
-    store.set('shikimoriCredentials', null)
-    store.set('shikimoriUser', null)
-    store.set('shikimoriUserRates', [])
-    store.set('shikimoriUpdateQueue', [])
-    store.set('shikimoriAnimeDetails', {})
-    prefetchAbort = true
-    invalidateCalendarCache()
-    stopSyncTimer()
-    broadcastToAll(EVENT_CHANNELS.SHIKIMORI_OFFLINE_QUEUE_CHANGED, { length: 0 })
-    broadcastSyncStatus()
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_USER, () => {
-    return store.get('shikimoriUser') as shikimori.ShikiUser | null
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_RATE, async (_event, malId: number) => {
-    const accessToken = await shikimori.ensureFreshToken(store)
-    const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-    if (!user) throw new Error('Not logged in to Shikimori')
-    return shikimori.getUserRate(accessToken, user.id, malId)
-  })
-
-  ipcMain.handle(
-    CHANNELS.SHIKIMORI_UPDATE_RATE,
-    async (
-      _event,
-      malId: number,
-      episodes: number,
-      status: shikimori.ShikiUserRateStatus,
-      score: number,
-      rewatches: number
-    ) => {
-      const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-      if (!user) throw new Error('Not logged in to Shikimori')
-
-      const cached = store.get('shikimoriUserRates') as {
-        rate: Record<string, unknown> & {
-          id?: number
-          target_id: number
-          episodes: number
-          status: shikimori.ShikiUserRateStatus
-          score: number
-          rewatches?: number
-        }
-        shikiAnime: unknown
-        smotretAnime: unknown
-      }[]
-      const idx = cached.findIndex((e) => e.rate.target_id === malId)
-      const prevStatus = idx !== -1 ? cached[idx].rate.status : undefined
-
-      try {
-        const accessToken = await shikimori.ensureFreshToken(store)
-        const existing = await shikimori.getUserRate(accessToken, user.id, malId)
-        const updatedRate = existing
-          ? await shikimori.updateUserRate(
-              accessToken,
-              existing.id,
-              episodes,
-              status,
-              score,
-              rewatches
-            )
-          : await shikimori.createUserRate(
-              accessToken,
-              user.id,
-              malId,
-              episodes,
-              status,
-              score,
-              rewatches
-            )
-
-        if (idx !== -1) {
-          cached[idx] = {
-            ...cached[idx],
-            rate: {
-              id: updatedRate.id,
-              score: updatedRate.score,
-              status: updatedRate.status,
-              episodes: updatedRate.episodes,
-              rewatches: updatedRate.rewatches ?? 0,
-              updated_at: new Date().toISOString(),
-              target_id: updatedRate.target_id
-            }
-          }
-          store.set('shikimoriUserRates', cached)
-          invalidateCalendarCache()
-          broadcastToAll(EVENT_CHANNELS.SHIKIMORI_RATE_UPDATED, cached[idx])
-          void maybeBroadcastCleanupPrompt(
-            cached[idx].smotretAnime,
-            malId,
-            updatedRate.status,
-            prevStatus
-          )
-        } else {
-          refreshShikimoriRatesInBackground()
-          void maybeBroadcastCleanupPrompt(null, malId, updatedRate.status, undefined)
-        }
-
-        void syncShikimoriQueue()
-        return updatedRate
-      } catch (err) {
-        if (!isNetworkError(err) || idx === -1) throw err
-
-        const cachedEntry = cached[idx]
-        const rateId = typeof cachedEntry.rate.id === 'number' ? cachedEntry.rate.id : null
-        const before = {
-          episodes: cachedEntry.rate.episodes,
-          status: cachedEntry.rate.status,
-          score: cachedEntry.rate.score,
-          rewatches: cachedEntry.rate.rewatches ?? 0
-        }
-        const after = { episodes, status, score, rewatches }
-
-        const queue = store.get('shikimoriUpdateQueue') as QueuedShikimoriUpdate[]
-        queue.push({ malId, rateId, before, after, queuedAt: Date.now() })
-        store.set('shikimoriUpdateQueue', queue)
-
-        cached[idx] = {
-          ...cachedEntry,
-          rate: {
-            ...cachedEntry.rate,
-            episodes,
-            status,
-            score,
-            rewatches,
-            updated_at: new Date().toISOString()
-          }
-        }
-        store.set('shikimoriUserRates', cached)
-        invalidateCalendarCache()
-        broadcastToAll(EVENT_CHANNELS.SHIKIMORI_RATE_UPDATED, cached[idx])
-        broadcastToAll(EVENT_CHANNELS.SHIKIMORI_OFFLINE_QUEUE_CHANGED, { length: queue.length })
-        startSyncTimer()
-        void maybeBroadcastCleanupPrompt(cached[idx].smotretAnime, malId, status, prevStatus)
-
-        return {
-          id: rateId ?? -1,
-          score,
-          status,
-          episodes,
-          rewatches,
-          target_id: malId,
-          target_type: 'Anime'
-        } satisfies shikimori.ShikiUserRate
-      }
-    }
-  )
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_FRIENDS_RATES, async (_event, malId: number) => {
-    const accessToken = await shikimori.ensureFreshToken(store)
-    const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-    if (!user) throw new Error('Not logged in to Shikimori')
-    return shikimori.getFriendsRatesForAnime(accessToken, user.id, malId)
-  })
-
-  async function fetchAndCacheShikimoriRates(status?: string): Promise<unknown[]> {
-    const accessToken = await shikimori.ensureFreshToken(store)
-    const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-    if (!user) throw new Error('Not logged in to Shikimori')
-    const rates = await shikimori.getUserAnimeRates(
-      accessToken,
-      user.id,
-      status as shikimori.ShikiUserRateStatus | undefined
-    )
-    const malIds = rates.map((r) => r.anime.id)
-    const malMap = await lookupByMalIds(malIds)
-    const entries = rates.map((rate) => ({
-      rate: {
-        id: rate.id,
-        score: rate.score,
-        status: rate.status,
-        episodes: rate.episodes,
-        rewatches: rate.rewatches ?? 0,
-        updated_at: rate.updated_at,
-        target_id: rate.target_id
-      },
-      shikiAnime: rate.anime,
-      smotretAnime: malMap[rate.anime.id] ?? null
-    }))
-    if (!status) {
-      store.set('shikimoriUserRates', entries)
-    }
-    return entries
-  }
-
-  function refreshShikimoriRatesInBackground(): void {
-    fetchAndCacheShikimoriRates()
-      .then((entries) => {
-        invalidateCalendarCache()
-        broadcastToAll(EVENT_CHANNELS.SHIKIMORI_RATES_REFRESHED, entries)
-        void syncShikimoriQueue()
-        void prefetchShikimoriDetails()
-        void runAutoDownloadTick('rates-refreshed').catch((err) =>
-          console.warn('[auto-dl] rates-refreshed tick failed:', err)
-        )
-      })
-      .catch((err) => console.warn('[shikimori] background refresh failed:', err))
-  }
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_ANIME_RATES, async (_event, status?: string) => {
-    if (!status) {
-      const cached = store.get('shikimoriUserRates') as unknown[]
-      if (cached.length > 0) {
-        refreshShikimoriRatesInBackground()
-        return cached
-      }
-    }
-    const entries = await fetchAndCacheShikimoriRates(status)
-    if (!status) void prefetchShikimoriDetails()
-    return entries
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_ANIME_DETAILS, (_event, malId: number) => {
-    const cache = store.get('shikimoriAnimeDetails') as Record<
-      string,
-      { details: shikimori.ShikiAnimeDetails; fetchedAt: number }
-    >
-    const entry = cache[String(malId)]
-    if (!entry) {
-      const creds = store.get('shikimoriCredentials') as shikimori.ShikiCredentials | null
-      if (creds) void prefetchShikimoriDetails()
-      return null
-    }
-    return entry.details
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_TRIGGER_DETAIL_PREFETCH, () => {
-    void prefetchShikimoriDetails()
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_OFFLINE_QUEUE_LENGTH, () => {
-    return (store.get('shikimoriUpdateQueue') as unknown[]).length
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_SYNC_STATUS, () => getSyncStatus())
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_TRIGGER_SYNC, () => {
-    void syncShikimoriQueue()
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_FRIENDS_ACTIVITY, async () => {
-    const accessToken = await shikimori.ensureFreshToken(store)
-    const user = store.get('shikimoriUser') as shikimori.ShikiUser | null
-    if (!user) throw new Error('Not logged in to Shikimori')
-
-    const activities = await shikimori.getFriendsActivity(accessToken, user.id)
-    const malIds = Array.from(new Set(activities.map((a) => a.malId)))
-    const malMap = await lookupByMalIds(malIds)
-
-    return activities.map((a) => ({
-      ...a,
-      smotretAnime: malMap[a.malId] ?? null
-    }))
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_CALENDAR, async (_event, force = false) => {
-    if (!force && calendarCache && Date.now() - calendarCache.fetchedAt < CALENDAR_CACHE_TTL_MS) {
-      return calendarCache.data
-    }
-
-    let cachedRates = store.get('shikimoriUserRates') as {
-      rate: { target_id?: number; episodes: number; status: shikimori.ShikiUserRateStatus }
-      shikiAnime?: { id: number }
-    }[]
-
-    if (cachedRates.length === 0) {
-      try {
-        await fetchAndCacheShikimoriRates()
-        cachedRates = store.get('shikimoriUserRates') as typeof cachedRates
-      } catch (err) {
-        console.warn('[shikimori] calendar: rate fetch failed:', err)
-      }
-    }
-
-    // Older cached entries may lack `rate.target_id`; fall back to `shikiAnime.id`
-    // (also the MAL ID) — same compatibility shim used by `home:get-continue-watching`.
-    function malIdOf(r: (typeof cachedRates)[number]): number | undefined {
-      return r.rate.target_id ?? r.shikiAnime?.id
-    }
-
-    const tracked = new Set<number>()
-    const statusByMal = new Map<number, shikimori.ShikiUserRateStatus>()
-    for (const r of cachedRates) {
-      const malId = malIdOf(r)
-      if (!malId) continue
-      statusByMal.set(malId, r.rate.status)
-      if (
-        r.rate.status === 'watching' ||
-        r.rate.status === 'rewatching' ||
-        r.rate.status === 'planned'
-      ) {
-        tracked.add(malId)
-      }
-    }
-
-    if (tracked.size === 0) {
-      calendarCache = { data: [], fetchedAt: Date.now() }
-      return []
-    }
-
-    const calendar = await shikimori.getCalendar()
-    const filtered = calendar.filter((c) => c.next_episode_at && tracked.has(c.anime.id))
-
-    const malIds = Array.from(new Set(filtered.map((c) => c.anime.id)))
-    const malMap = await lookupByMalIds(malIds)
-
-    const SHIKI_BASE = 'https://shikimori.one'
-    function absoluteImage(url: string): string {
-      return url.startsWith('http') ? url : `${SHIKI_BASE}${url}`
-    }
-
-    const entries: CalendarEntry[] = filtered.map((c) => {
-      const malId = c.anime.id
-      const smotret = malMap[malId] ?? null
-      const posterUrl =
-        smotret?.posterUrlSmall ||
-        absoluteImage(c.anime.image.preview || c.anime.image.x96 || c.anime.image.original)
-      return {
-        malId,
-        animeId: smotret?.id ?? null,
-        name: c.anime.russian || c.anime.name,
-        posterUrl,
-        kind: c.anime.kind,
-        episodeInt: String(c.next_episode ?? c.anime.episodes_aired + 1),
-        nextEpisodeAt: c.next_episode_at!,
-        userStatus: statusByMal.get(malId) ?? 'planned'
-      }
-    })
-
-    calendarCache = { data: entries, fetchedAt: Date.now() }
-    return entries
-  })
-
-  ipcMain.handle(CHANNELS.SHIKIMORI_GET_RELATED, async (_event, malId: number) => {
-    const franchise = await shikimori.getFranchise(malId)
-
-    const CANONICAL_RELATIONS = new Set([
-      'sequel',
-      'prequel',
-      'side_story',
-      'parent_story',
-      'alternative_version',
-      'summary',
-      'full_story',
-      'spin_off',
-      'alternative_setting'
-    ])
-
-    const adjacency = new Map<number, { nodeId: number; relation: string }[]>()
-    for (const link of franchise.links) {
-      if (!CANONICAL_RELATIONS.has(link.relation)) continue
-      if (!adjacency.has(link.source_id)) adjacency.set(link.source_id, [])
-      adjacency.get(link.source_id)!.push({ nodeId: link.target_id, relation: link.relation })
-      if (!adjacency.has(link.target_id)) adjacency.set(link.target_id, [])
-      adjacency.get(link.target_id)!.push({ nodeId: link.source_id, relation: link.relation })
-    }
-
-    const reachable = new Set<number>([franchise.current_id])
-    const queue = [franchise.current_id]
-    for (let i = 0; i < queue.length; i++) {
-      const cur = queue[i]
-      for (const { nodeId } of adjacency.get(cur) || []) {
-        if (reachable.has(nodeId)) continue
-        reachable.add(nodeId)
-        queue.push(nodeId)
-      }
-    }
-
-    const filteredNodes = franchise.nodes.filter((n) => reachable.has(n.id))
-    const malIds = Array.from(new Set(filteredNodes.map((n) => n.id)))
-    const malMap = await lookupByMalIds(malIds)
-
-    const relationByNodeId = new Map<number, string>()
-    for (const link of franchise.links) {
-      if (link.source_id === franchise.current_id) {
-        relationByNodeId.set(link.target_id, link.relation)
-      }
-    }
-
-    const cachedDetails = store.get('shikimoriAnimeDetails') as Record<
-      string,
-      { details: shikimori.ShikiAnimeDetails; fetchedAt: number }
-    >
-    const cachedRates = store.get('shikimoriUserRates') as {
-      rate: { status: shikimori.ShikiUserRateStatus }
-      shikiAnime: { id: number; kind?: string }
-    }[]
-    const kindByMalId = new Map<number, string>()
-    const statusByMalId = new Map<number, shikimori.ShikiUserRateStatus>()
-    for (const r of cachedRates) {
-      const id = r.shikiAnime?.id
-      if (id == null) continue
-      if (r.shikiAnime.kind) kindByMalId.set(id, r.shikiAnime.kind)
-      if (r.rate?.status) statusByMalId.set(id, r.rate.status)
-    }
-
-    function normalizeKind(
-      node: shikimori.ShikiFranchiseNode,
-      smotretType: string | undefined
-    ): string | null {
-      const cached = cachedDetails[String(node.id)]?.details?.kind
-      if (cached) return cached
-      const rated = kindByMalId.get(node.id)
-      if (rated) return rated
-      if (smotretType) return smotretType.toLowerCase()
-      const k = (node.kind || '').toLowerCase()
-      if (k.includes('проморолик') || k === 'pv') return 'pv'
-      if (k.includes('реклама') || k === 'cm') return 'cm'
-      if (k.includes('клип') || k === 'music') return 'music'
-      if (k.includes('тв-спешл') || k === 'tv_special') return 'tv_special'
-      if (k.includes('спешл') || k === 'special') return 'special'
-      if (k.includes('полнометраж') || k === 'movie') return 'movie'
-      if (k.includes('ova')) return 'ova'
-      if (k.includes('ona')) return 'ona'
-      if (k.includes('tv')) return 'tv'
-      return node.kind ?? null
-    }
-
-    const EXCLUDED_KINDS = new Set(['pv', 'cm', 'music'])
-
-    const sorted = [...filteredNodes].sort((a, b) => {
-      if (a.date == null && b.date == null) return 0
-      if (a.date == null) return 1
-      if (b.date == null) return -1
-      return a.date - b.date
-    })
-
-    return sorted
-      .map((n) => {
-        const smotret = malMap[n.id] ?? null
-        const kind = normalizeKind(n, smotret?.type)
-        return {
-          relation: relationByNodeId.get(n.id) ?? null,
-          shikiAnime: {
-            id: n.id,
-            name: n.name,
-            image_url: n.image_url,
-            url: n.url,
-            year: n.year,
-            kind,
-            date: n.date
-          },
-          smotretAnime: smotret,
-          isCurrent: n.id === franchise.current_id,
-          watchStatus: statusByMalId.get(n.id) ?? null
-        }
-      })
-      .filter((e) => e.isCurrent || !(e.shikiAnime.kind && EXCLUDED_KINDS.has(e.shikiAnime.kind)))
   })
 
   ipcMain.handle(CHANNELS.SHELL_OPEN_EXTERNAL, async (_event, url: string) => {
@@ -3056,7 +2251,7 @@ async function bootstrap(): Promise<void> {
     downloadManager,
     broadcast: broadcastToAll,
     isShikimoriLoggedIn: () => Boolean(store.get('shikimoriCredentials')),
-    refreshShikimoriDetails: refreshShikimoriDetailsForMalId
+    refreshShikimoriDetails: (malId) => shikimoriSyncService.refreshShikimoriDetailsForMalId(malId)
   })
   startAutoDownloaderTimer()
   setTimeout(() => {
@@ -3117,9 +2312,9 @@ async function bootstrap(): Promise<void> {
     broadcastToAll(EVENT_CHANNELS.SYNCPLAY_TRACE, entry)
   })
 
-  if (getQueueLength() > 0) {
-    startSyncTimer()
-    void syncShikimoriQueue()
+  if (shikimoriSyncService.getQueueLength() > 0) {
+    shikimoriSyncService.startSyncTimer()
+    void shikimoriSyncService.syncShikimoriQueue()
   }
 
   app.on('activate', () => {
